requirements.txt
certifi==2018.8.13
chardet==3.0.4
Django>=2.1.2
django-cors-headers==2.4.0
djangorestframework==3.8.2
docopt==0.6.2
idna==2.7
nltk==3.3
pipenv==2018.7.1
pipreqs==0.4.9
pytz==2018.5
requests>=2.20.0
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
yarg==0.1.9

